requirements.txt
# RentenDashboard Requirements

# Core HTTP and OAuth libraries
requests>=2.31.0
requests-oauthlib>=1.3.1
authlib>=1.2.1

# Async HTTP client
aiohttp>=3.8.5
aiofiles>=23.1.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Cryptography for secure token handling
cryptography>=41.0.0

# Date/time handling
python-dateutil>=2.8.2

# Logging and monitoring
structlog>=23.1.0

# Testing (development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development tools
black>=23.3.0
flake8>=6.0.0
mypy>=1.4.0

# Optional: Rich for enhanced console output
rich>=13.4.0